+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- [#12] HTTP head as route annotation. via [#27]
 - [#9] Possibility to define multiple validators. via [#24]
 - Documentation
 - Code coverage reporting. via [#26]
@@ -31,7 +32,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [#17]: https://github.com/smartive/giuseppe/issues/17
 [#18]: https://github.com/smartive/giuseppe/issues/18
 [#9]: https://github.com/smartive/giuseppe/issues/9
+[#17]: https://github.com/smartive/giuseppe/issues/12
 
 [#21]: https://github.com/smartive/giuseppe/pull/21
 [#24]: https://github.com/smartive/giuseppe/pull/24
 [#26]: https://github.com/smartive/giuseppe/pull/26
+[#27]: https://github.com/smartive/giuseppe/pull/27